requirements.txt
configparser-5.3.0.dist-info
customtkinter-5.1.2.dist-info
setuptools-65.5.0.dist-info